+++ CHANGELOG.md
@@ -7,8 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
+- Function argument help in DefinitionEditor, based on cursor position.
 - ComplarePlot has added generations list in Dialog, with editable labels.
-- Add autocomplete to DefinitionDialog (suggests functions and linked primitives).
+- Autocomplete to DefinitionDialog (suggests functions and linked primitives).
 - CHANGELOG.
 - Progresive Web App with chromium based browsers.
 - KeepResults option saved to file for ComparePlots.